AP_InertialSensor: add fix for icm42688 stuck gyro issue
The ICM42688 has a feature called AFSR where it automatically switches from a low noise, low range to higher noise, high range mode if needed. During this switch, the gyro data output may freeze for a few milliseconds, leading to unwanted behaviour. This code is to disable the AFSR feature on IMU init.